The Code: (new to python) - Need Help
why does the if statement under '#help here" allow thing_to_add to be added to my_list? the else: statement should be the condition to do so, not the if statement, in turn, i need to remove what is being added in the if statement, but that messes up the order of the list, anyone have any idea?